requirements.txt
daphne
django
djangorestframework
djangorestframework_simplejwt
channels
channels_redis
pyjwt